Function demo_src::handle_full_data_option

fn handle_full_data_option(
    src: &mut SacnSource, 
    split_input: Vec<&str>
) -> Result<bool>

Handles the user command to send a full universe of data which starts with the data given and then is padded with 0's upto the full length.

Arguments

src: A mutable reference to the SacnSource to use to send the data with.

split_input: The parts of the user command which have been split up by white space.

split_input[1] is expected to be the universe.

split_input[2] is expected to be the syncronisation universe.

split_input[3] is expected to be the priority.

split_input[4] is expected to be the start of the data to send.