Function demo_src::run_acceptance_test_demo

fn run_acceptance_test_demo(src: &mut SacnSource) -> Result<()>

Runs the acceptance test sender to vision visualiser demo.

Made to work with the corresponding vision visualiser "Student-Union-Model.v3s" file with patch as follows: Format: <fixture_name> <channel_count>ch: <sACN_universe>-, ....

Patch: Robe Robin LedBeam 150 16ch: 1-1, 1-50, 1-100, 1-150, 1-200, 1-250, 1-300, 1-350. Fresnel-Front-Light 1ch: 2-1, 2-2, 2-3

Step 1, 150 + Front On at full. Step 2, Red Step 3, Blue Step 4, All off

Arguments

src: A mutable reference to the SacnSource to use as the sender in the acceptance test.