Function demo_src::run_test_2_universes_distinct_values

fn run_test_2_universes_distinct_values(
    src: &mut SacnSource, 
    uni_1: u16, 
    uni_2: u16, 
    uni1_val: u8, 
    uni2_val: u8, 
    dst_ip: Option<SocketAddr>
) -> Result<()>

Constantly sends data packets to 2 universes with the given values. Used as part of the interoperability testing as described in the the Interoperability Testing document.

Arguments:

src: A mutable reference to the SacnSource to use as the sender in this test.

uni_1: The first universe to send data on.

uni_2: The second universe to send data on.

uni1_val: The value to send on the first universe.

uni2_val: The value to send on the second universe.

dst_ip: None to use multicast or Some(addr) to use unicast to a specific address.